+++ packages/ethereum-contracts/contracts/agreements/ConstantFlowAgreementV1.sol
@@ -453,6 +453,8 @@ contract ConstantFlowAgreementV1 is
                 ctx, currentContext);
         }
 
+        flowVars.token.emitPseudoTransfer(flowVars.sender, flowVars.receiver);
+
         _requireAvailableBalance(flowVars.token, flowVars.sender, currentContext);
     }
 
@@ -592,6 +594,8 @@ contract ConstantFlowAgreementV1 is
                     newCtx, currentContext);
             }
         }
+
+        flowVars.token.emitPseudoTransfer(flowVars.sender, flowVars.receiver);
     }
 
     /**************************************************************************