+++ packages/ethereum-contracts/contracts/agreements/gdav1/SuperfluidPool.sol
@@ -453,6 +453,11 @@ contract SuperfluidPool is ISuperfluidPool, BeaconProxiable {
             _membersData[memberAddr] = _pdPoolMemberToMemberData(pdPoolMember, claimedValue);
             assert(GDA.appendIndexUpdateByPool(superToken, p, t));
         }
+
+        if ((oldUnits == 0 || newUnits == 0) && oldUnits != newUnits) {
+            GDA.tokenEmitPseudoTransfer(superToken, address(this), memberAddr);
+        }
+
         emit MemberUnitsUpdated(superToken, memberAddr, oldUnits, newUnits);
     }