+++ packages/ethereum-contracts/contracts/superfluid/SuperfluidToken.sol
@@ -376,6 +376,10 @@ abstract contract SuperfluidToken is ISuperfluidToken
         );
     }
 
+    function emitPseudoTransfer(address from, address to) external onlyAgreement {
+        emit IERC20.Transfer(from, to, 0);
+    }
+
     /**************************************************************************
     * Modifiers
     *************************************************************************/