+++ packages/ethereum-contracts/contracts/utils/ERC1820Implementer.sol
@@ -3,7 +3,6 @@ pragma solidity ^0.8.23;
 
 import { IERC1820Implementer } from "@openzeppelin-v5/contracts/interfaces/IERC1820Implementer.sol";
 
-
 // Copy of OpenZeppelin v4.9.6: https://github.com/OpenZeppelin/openzeppelin-contracts/blob/v4.9.6/
 /**
  * @dev Implementation of the {IERC1820Implementer} interface.